Move utility methods into TreePathUtil
