+++ annotation-file-utilities/src/annotator/find/IsCriterion.java
@@ -1,6 +1,6 @@
 package annotator.find;
 
-import annotator.scanner.CommonScanner;
+import annotator.scanner.TreePathUtil;
 import com.sun.source.tree.*;
 import com.sun.source.util.TreePath;
 
@@ -35,7 +35,7 @@ final class IsCriterion implements Criterion {
       return false;
     }
     Tree tree = path.getLeaf();
-    if (CommonScanner.hasClassKind(tree)) {
+    if (TreePathUtil.hasClassKind(tree)) {
       return InClassCriterion.isSatisfiedBy(path, name, /*exactMatch=*/ true);
     }
     if (tree.getKind() != kind) {
